requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
auth0-python==4.7.2
autobahn==24.4.2
Automat==24.8.1
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.35.22
botocore==1.35.22
build==1.2.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
cleo==2.1.0
click==8.1.7
cohere==5.9.2
coloredlogs==15.0.1
comm==0.2.2
constantly==23.10.4
crashtest==0.4.1
crewai==0.98.0
cryptography==43.0.1
daphne==4.1.2
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
diffusers==0.30.0
distlib==0.3.8
distro==1.9.0
Django==5.0.2
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
dulwich==0.21.7
embedchain==0.1.125
et_xmlfile==2.0.0
evolutionapi==0.0.4
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.5
fastavro==1.9.7
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.2
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.10.0
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-aiplatform==1.67.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-genai==0.3.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
gptcache==0.1.44
gradio==5.6.0
gradio_client==1.4.3
greenlet==3.0.3
groq==0.11.0
grpc-google-iam-v1==0.13.1
grpcio==1.65.5
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.0.0
importlib_resources==6.4.5
incremental==24.7.2
iniconfig==2.0.0
installer==0.7.0
instructor==1.3.3
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.25.3
jsonpatch==1.33
jsonpickle==4.0.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==24.3.1
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.3.7
langchain-anthropic==0.1.23
langchain-cohere==0.3.1
langchain-community==0.3.7
langchain-core==0.3.19
langchain-experimental==0.3.3
langchain-groq==0.1.9
langchain-huggingface==0.1.0
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langsmith==0.1.143
limits==4.0.0
litellm==1.57.4
load-dotenv==0.1.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.29
mmh3==4.1.0
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.9.1
numpy==2.2.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.20.0
openai==1.60.1
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
ordered-set==4.1.0
orjson==3.10.11
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.11.1
platformdirs==4.2.2
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
portalocker==2.10.1
posthog==3.6.6
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.5
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.6.1
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.0
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.9.18
pyOpenSSL==25.0.0
pypdf==5.1.0
PyPDF2==3.0.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyright==1.1.381
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
python-telegram-bot==21.4
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyVISA==1.14.1
PyYAML==6.0.1
pyzmq==26.2.0
qdrant-client==1.11.0
rapidfuzz==3.9.6
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
ruff==0.7.4
s3transfer==0.10.2
safehttpx==0.1.1
safetensors==0.4.5
schema==0.7.7
scikit-learn==1.5.1
SecretStorage==3.3.3
selenium==4.23.1
semantic-version==2.10.0
semver==3.0.2
sentence-transformers==3.0.1
service-identity==24.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
SQLAlchemy-Utils==0.41.2
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.1.0
tomli_w==1.1.0
tomlkit==0.12.0
torch==2.4.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.44.2
trio==0.26.2
trio-websocket==0.11.1
triton==3.0.0
trove-classifiers==2024.7.2
Twisted==24.11.0
txaio==23.1.1
typer==0.13.1
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.3
uv==0.5.2
uvicorn==0.32.0
uvloop==0.20.0
virtualenv==20.26.3
watchfiles==0.23.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.11.1
zipp==3.20.2
zope.interface==7.2
django-environ==0.11.2
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
pyvisa-py==0.7.1
pyserial==3.5
uritemplate==4.1.1
